// frontend/src/types/expertComment.ts
export interface ExpertComment {
  id: number;
  comment: string;
  comment_assess_vital_signs_destination: string;
  other_text_providers_impressions: string;
  text: string;
  text_trend_defib_notes: string;
  text_objectiv_circulation_others: string;
  text_objectiv_circulations: string;
  text_objectiv_skin_assessment: string;
  text_pri_assessment_circ_others: string;
  text_pri_assessment_circulation: string;
  text_pri_assessment_skin: string;
  type_providers_impressions_open: number;
  type_providers_impressions_closed: number;
  stable_vitals: number;
  vital_measurement_not_necessary: number;
  patient_transported_in_ambulance: number;
  bleeding_lower_torso: number;
  bleeding_neck: number;
  bleeding_upper_extremities: number;
  bleeding_lower_extremities: number;
  bleeding_other: number;
  patient_transferred: number;
  ruled_out_acute_coronary_syndrome: number;
  ruled_out_stroke: number;
  ruled_out_spinal_disc_herniation: number;
  ruled_out_fracture: number;
  ruled_out_concussion: number;
  pain_noted: number;
  fainting_noted: number;
  bleeding_noted: number;
  worsening_condition_noted: number;
  fracture_noted: number;
  infection_noted: number;
  dizziness_noted: number;
  joint_locking_noted: number;
  nausea_noted: number;
  morphine_administered: number;
  epinephrine_administered: number;
  bleeding_upper_torso: number;
  catecholamine_required: number;
  weak_pulse: number;
  strong_pulse: number;
  normal_skin: number;
  pale_skin: number;
  decreased_skin_turgor: number;
  sweaty_skin: number;
  cold_skin: number;
  warm_skin: number;
  acute_coronary_syndrome_noted: number;
  heart_issue_noted: number;
  normal_table_result: number;
  abnormal_table_result: number;
  severe_table_result: number;
}